app/requirements.txt
Flask==2.3.3
pymysql==1.1.0
mysql-connector-python==8.3.0
werkzeug==2.3.7
python-dotenv==1.0.1